Add files via upload
This is version 1.1 of the GAHT provider database. It contains only the name of facility, street address, city, state, ZIP code and lat/long of each facility. It's vector data (points) in GeoJSON format and can be added directly to a GIS (like QGIS, GeoDa, or the ESRI suite). By default it is not projected because it can be disaggregated by state to measure local availability. It's designed for research purposes first and foremost and was not designed to be user-facing or as a tool for trans people seeking GAHT to find care providers. These data are pulled from existing sources, which are linked below, as well as from web searches and snowballing. Inclusion in the database is not an endorsement of services provided; rather, inclusion is merely an acknowledgement that GAHT is provided by one or more providers at a given facility according to publicly available information. Data were collected between Fall 2019 and Spring 2021 and have not been updated since.
https://www.reddit.com/r/transgender/comments/d76pux/erins_informed_consent_hrt_map_of_the_us_google/
https://www.outcarehealth.org/resources/
https://resource.genderinfinity.org/
https://southernequality.org/resources/transinthesouth/